# Very minor performance nit in AbstractAction constructor