--- a/inc/Action/AbstractAction.php
+++ b/inc/Action/AbstractAction.php
@@ -28,8 +28,7 @@ abstract class AbstractAction
             $this->actionname = $actionname;
         } else {
             // As of PHP 8 this seems to be the fastest way to get the name:
-            $n = explode('\\', get_class($this));
-            $this->actionname = strtolower(array_pop($n));
+            $this->actionname = strtolower((new \ReflectionClass($this))->getShortName());
         }
     }