@splitbrain 2024-08-06
I'm confused, when sorting your results by speed, reflection is fastest. Explode is slowest. But your PR introduces explode?

```
Reflection:     0.007136058807373 s ClassA
Substring:      0.0079569101333618 s ClassA
SubstringStrCh: 0.0082878351211548 s ClassA
Basename:       0.01236629486084 s ClassA
Explode:        0.012751793861389 s ClassA
```